Enum move_cli::sandbox::cli::SandboxCommand

pub enum SandboxCommand {
    Publish {
        source_files: Vec<String>,
        no_republish: bool,
        ignore_breaking_changes: bool,
        override_ordering: Option<Vec<String>>,
    },
    Run {
        script_file: PathBuf,
        script_name: Option<String>,
        signers: Vec<String>,
        args: Vec<TransactionArgument>,
        type_args: Vec<TypeTag>,
        gas_budget: Option<u64>,
        dry_run: bool,
    },
    Test {
        path: PathBuf,
        use_temp_dir: bool,
        track_cov: bool,
        create: bool,
    },
    View {
        file: PathBuf,
    },
    Clean {},
    Doctor {},
    Link {
        source_files: Vec<String>,
        no_republish: bool,
    },
    Generate {
        cmd: GenerateCommand,
    },
}

Variants

Publish

Compile the specified modules and publish the resulting bytecodes in global storage.

Fields of Publish

source_files: Vec<String>

The source files containing modules to publish.

no_republish: bool

If set, fail during compilation when attempting to publish a module that already exists in global storage.

ignore_breaking_changes: bool

By default, code that might cause breaking changes for bytecode linking or data layout compatibility checks will not be published. Set this flag to ignore breaking changes checks and publish anyway.

override_ordering: Option<Vec<String>>

Manually specify the publishing order of modules.

Run

Compile/run a Move script that reads/writes resources stored on disk in storage. This command compiles the script first before running it.

Fields of Run

script_file: PathBuf

Path to .mv file containing either script or module bytecodes. If the file is a module, the script_name parameter must be set.

script_name: Option<String>

Name of the script function inside script_file to call. Should only be set if script_file points to a module.

signers: Vec<String>

Possibly-empty list of signers for the current transaction (e.g., account in main(&account: signer)). Must match the number of signers expected by script_file.

args: Vec<TransactionArgument>

Possibly-empty list of arguments passed to the transaction (e.g., i in main(i: u64)). Must match the arguments types expected by script_file. Supported argument types are bool literals (true, false), u64 literals (e.g., 10, 58), address literals (e.g., 0x12, 0x0000000000000000000000000000000f), hexadecimal strings (e.g., x“0012“ will parse as the vector value [00, 12]), and ASCII strings (e.g., ’b“hi“ will parse as the vector value [68, 69]).

type_args: Vec<TypeTag>

Possibly-empty list of type arguments passed to the transaction (e.g., T in main<T>()). Must match the type arguments kinds expected by script_file.

gas_budget: Option<u64>

Maximum number of gas units to be consumed by execution. When the budget is exhaused, execution will abort. By default, no gas-budget is specified and gas metering is disabled.

dry_run: bool

If set, the effects of executing script_file (i.e., published, updated, and deleted resources) will NOT be committed to disk.

Test

Run expected value tests using the given batch file.

Fields of Test

path: PathBuf

a directory path in which all the tests will be executed.

use_temp_dir: bool

Use an ephemeral directory to serve as the testing workspace. By default, the directory containing the args.txt will be the workspace.

track_cov: bool

Show coverage information after tests are done. By default, coverage will not be tracked nor shown.

create: bool

Create a new test directory scaffold with the specified .

View

View Move resources, events files, and modules stored on disk.

Fields of View

file: PathBuf

Path to a resource, events file, or module stored on disk.

Clean

Delete all resources, events, and modules stored on disk under storage. Does not delete anything in src.

Fields of Clean

Doctor

Run well-formedness checks on the storage and build directories.

Fields of Doctor

Link

Typecheck and verify the scripts and/or modules under src.

Fields of Link

source_files: Vec<String>

The source files containing modules to publish.

no_republish: bool

If set, fail when attempting to typecheck a module that already exists in global storage.

Generate

Generate struct layout bindings for the modules stored on disk under storage

Fields of Generate

cmd: GenerateCommand

Implementations

impl SandboxCommand

pub fn handle_command(
&self,
natives: Vec<(AccountAddress, Identifier, Identifier, NativeFunction)>,
error_descriptions: &ErrorMapping,
move_args: &Move,
mode: &Mode
) -> Result<()>