Struct move_cli::OnDiskStateView

pub struct OnDiskStateView { /* fields omitted */ }

Implementations

impl OnDiskStateView

pub fn create(
move_data_dir: PathBuf,
compiled_modules: &[CompiledModule]
) -> Result<Self>

Create an OnDiskStateView that reads/writes resource data in move_data_dir and can execute code in compiled_modules.

pub fn is_resource_path(&self, p: &Path) -> bool

pub fn is_event_path(&self, p: &Path) -> bool

pub fn is_module_path(&self, p: &Path) -> bool

pub fn get_event_path(&self, key: &EventKey) -> PathBuf

pub fn get_resource_bytes(
&self,
addr: AccountAddress,
tag: StructTag
) -> Result<Option<Vec<u8>>>

Read the resource bytes stored on-disk at addr/tag

pub fn get_compiled_module(
&self,
module_id: &ModuleId
) -> Result<CompiledModule>

Deserialize and return the module stored on-disk at addr/module_id

pub fn resolve_function(
&self,
module_id: &ModuleId,
idx: u16
) -> Result<Identifier>

Return the name of the function at idx in module_id

pub fn view_resource(
&self,
resource_path: &Path
) -> Result<Option<AnnotatedMoveStruct>>

pub fn view_events(&self, events_path: &Path) -> Result<Vec<AnnotatedMoveValue>>

pub fn view_module(&self, module_path: &Path) -> Result<Option<String>>

pub fn delete_resource(
&self,
addr: AccountAddress,
tag: StructTag
) -> Result<()>

Delete resource stored on disk at the path addr/tag

pub fn save_resource(
&self,
addr: AccountAddress,
tag: StructTag,
layout: MoveTypeLayout,
resource: Value
) -> Result<()>

Save resource on disk under the path addr/tag

pub fn save_event(
&self,
event_key: &[u8],
event_sequence_number: u64,
event_type: TypeTag,
event_layout: &MoveTypeLayout,
event_value: Value
) -> Result<()>

pub fn save_modules(&self) -> Result<bool>

Save all the modules in the local cache Returns true if any modules were saved