Crate move_cli

Modules

base

experimental

sandbox

Structs

Move

MoveCLI
MoveCLI is the CLI that will be executed by the move-cli command The cmd argument is added here rather than in Move to make it easier for other crates to extend move-cli

Enums

Command

ConcretizeMode

ExperimentalCommand

SandboxCommand

Constants

DEFAULT_BUILD_DIR
Default directory for build output

DEFAULT_DEP_MODE
Default dependency inclusion mode

DEFAULT_PACKAGE_DIR
Default directory where Move packages live under build_dir

DEFAULT_SOURCE_DIR
Default directory where Move modules live

DEFAULT_STORAGE_DIR
Default directory where saved Move resources live

Functions

move_cli

run_cli