Module move_cli::sandbox::commands::test

Constants

TEST_ARGS_FILENAME
The filename that contains the arguments to the Move binary.

Functions

create_test_scaffold
Create a directory scaffold for writing a Move CLI test.

run_all

run_one
Run the args_path batch file withcli_binary