Struct move_cli::OnDiskStateView

pub struct OnDiskStateView { /* fields omitted */ }

Implementations

impl OnDiskStateView

pub fn create<P: Into<PathBuf>>(build_dir: P, storage_dir: P) -> Result<Self>

Create an OnDiskStateView that reads/writes resource data and modules in storage_dir.

pub fn interface_files_dir(&self) -> Result<String>

pub fn is_resource_path(&self, p: &Path) -> bool

pub fn is_event_path(&self, p: &Path) -> bool

pub fn is_module_path(&self, p: &Path) -> bool

pub fn get_resource_bytes(
&self,
addr: AccountAddress,
tag: StructTag
) -> Result<Option<Vec<u8>>>

Read the resource bytes stored on-disk at addr/tag

pub fn has_module(&self, module_id: &ModuleId) -> bool

Check if a module at addr/module_id exists

pub fn get_compiled_module(
&self,
module_id: &ModuleId
) -> Result<CompiledModule>

Deserialize and return the module stored on-disk at addr/module_id

pub fn resolve_function(
&self,
module_id: &ModuleId,
idx: u16
) -> Result<Identifier>

Return the name of the function at idx in module_id

pub fn view_resource(
&self,
resource_path: &Path
) -> Result<Option<AnnotatedMoveStruct>>

Returns a deserialized representation of the resource value stored at resource_path. Returns Err if the path does not hold a resource value or the resource cannot be deserialized

pub fn view_events(&self, events_path: &Path) -> Result<Vec<AnnotatedMoveValue>>

pub fn view_module(&self, module_path: &Path) -> Result<Option<String>>

pub fn delete_resource(
&self,
addr: AccountAddress,
tag: StructTag
) -> Result<()>

Delete resource stored on disk at the path addr/tag

pub fn save_resource(
&self,
addr: AccountAddress,
tag: StructTag,
layout: MoveTypeLayout,
resource: Value
) -> Result<()>

Save resource on disk under the path addr/tag

pub fn save_resource_bytes(
&self,
addr: AccountAddress,
tag: StructTag,
bcs_bytes: &[u8]
) -> Result<()>

pub fn save_event(
&self,
event_key: &[u8],
event_sequence_number: u64,
event_type: TypeTag,
event_layout: &MoveTypeLayout,
event_value: Value
) -> Result<()>

pub fn save_contract_event(&self, event: ContractEvent) -> Result<()>

pub fn save_module(
&self,
module_id: &ModuleId,
module_bytes: &[u8]
) -> Result<()>

Save module on disk under the path module.address()/module.name()

pub fn save_modules(&self, modules: &[(ModuleId, Vec<u8>)]) -> Result<()>

Save all the modules in the local cache, re-generate mv_interfaces if required.

pub fn delete_module(&self, id: &ModuleId) -> Result<()>

pub fn resource_paths(&self) -> impl Iterator<Item = PathBuf> + '_

pub fn module_paths(&self) -> impl Iterator<Item = PathBuf> + '_

pub fn event_paths(&self) -> impl Iterator<Item = PathBuf> + '_

pub fn get_all_modules(&self) -> Result<BTreeMap<ModuleId, CompiledModule>>

Return a map of module ID -> module for all modules in the self.storage_dir. Returns an Err if a module does not deserialize